+++ vtm-android-example/src/org/oscim/android/test/RotateMarkerOverlayActivity.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2014 Hannes Janetzek
- * Copyright 2016 devemux86
+ * Copyright 2016-2017 devemux86
  * Copyright 2017 Longri
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
@@ -18,111 +18,35 @@
  */
 package org.oscim.android.test;
 
-import android.graphics.drawable.Drawable;
-import android.os.Bundle;
 import android.widget.Toast;
 
-import org.oscim.backend.canvas.Bitmap;
-import org.oscim.core.GeoPoint;
-import org.oscim.event.Gesture;
-import org.oscim.event.GestureListener;
-import org.oscim.event.MotionEvent;
-import org.oscim.layers.Layer;
-import org.oscim.layers.TileGridLayer;
-import org.oscim.layers.marker.ItemizedLayer;
 import org.oscim.layers.marker.MarkerItem;
-import org.oscim.layers.marker.MarkerSymbol;
-import org.oscim.layers.marker.MarkerSymbol.HotspotPlace;
-import org.oscim.map.Map;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import static org.oscim.android.canvas.AndroidGraphics.drawableToBitmap;
-import static org.oscim.tiling.source.bitmap.DefaultSources.STAMEN_TONER;
+public class RotateMarkerOverlayActivity extends MarkerOverlayActivity {
 
-public class RotateMarkerOverlayActivity extends BitmapTileMapActivity
-        implements ItemizedLayer.OnItemGestureListener<MarkerItem> {
+    private Timer timer;
+    private int markerCount = 0;
 
-    protected static final boolean BILLBOARDS = true;
-    protected MarkerSymbol mFocusMarker;
-    private ItemizedLayer<MarkerItem> markerLayer;
-
-    public RotateMarkerOverlayActivity() {
-        super(STAMEN_TONER.build());
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        mBitmapLayer.tileRenderer().setBitmapAlpha(0.5f);
-
-        // Map events receiver
-        mMap.layers().add(new MapEventsReceiver(mMap));
-
-        /* directly load bitmap from resources */
-        Bitmap bitmap = drawableToBitmap(getResources(), R.drawable.marker_poi);
-
-        MarkerSymbol symbol;
-        if (BILLBOARDS)
-            symbol = new MarkerSymbol(bitmap, HotspotPlace.BOTTOM_CENTER);
-        else
-            symbol = new MarkerSymbol(bitmap, HotspotPlace.CENTER, false);
-
-        /* another option: use some bitmap drawable */
-        Drawable d = getResources().getDrawable(R.drawable.marker_focus);
-        if (BILLBOARDS)
-            mFocusMarker = new MarkerSymbol(drawableToBitmap(d), HotspotPlace.BOTTOM_CENTER);
-        else
-            mFocusMarker = new MarkerSymbol(drawableToBitmap(d), HotspotPlace.CENTER, false);
-
-        markerLayer =
-                new ItemizedLayer<>(mMap, new ArrayList<MarkerItem>(),
-                        symbol, this);
-
-        mMap.layers().add(markerLayer);
-
-        List<MarkerItem> pts = new ArrayList<>();
-
-        for (double lat = -90; lat <= 90; lat += 5) {
-            for (double lon = -180; lon <= 180; lon += 5)
-                pts.add(new MarkerItem(lat + "/" + lon, "", new GeoPoint(lat, lon)));
-        }
-
-        markerLayer.addItems(pts);
-
-        mMap.layers().add(new TileGridLayer(mMap, getResources().getDisplayMetrics().density));
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-
-        /* ignore saved position */
-        mMap.setMapPosition(0, 0, 1 << 2);
-    }
-
-    Timer timer;
-    int markerCount = 0;
-
-    @Override
-    public boolean onItemSingleTapUp(int index, final MarkerItem item) {
+    private void itemEvent(final MarkerItem item) {
         if (item.getMarker() == null) {
             item.setMarker(mFocusMarker);
             markerCount++;
             final AtomicInteger rotValue = new AtomicInteger(0);
-            if (timer != null) timer.cancel();
+            if (timer != null)
+                timer.cancel();
             timer = new Timer();
             TimerTask timerTask = new TimerTask() {
                 @Override
                 public void run() {
                     float value = (float) (rotValue.incrementAndGet() * 10);
                     item.setRotation(value);
-                    if (rotValue.get() > 36) rotValue.set(0);
-                    markerLayer.update();
+                    if (rotValue.get() > 36)
+                        rotValue.set(0);
+                    mMarkerLayer.update();
                     mMap.updateMap(true);
                 }
             };
@@ -131,61 +55,24 @@ public class RotateMarkerOverlayActivity extends BitmapTileMapActivity
         } else {
             item.setMarker(null);
             markerCount--;
-            if (timer != null && markerCount == 0) timer.cancel();
+            if (timer != null && markerCount == 0)
+                timer.cancel();
         }
+    }
+
+    @Override
+    public boolean onItemSingleTapUp(int index, MarkerItem item) {
+        itemEvent(item);
 
         Toast.makeText(this, "Marker tap\n" + item.getTitle(), Toast.LENGTH_SHORT).show();
         return true;
     }
 
     @Override
-    public boolean onItemLongPress(int index, final MarkerItem item) {
-        if (item.getMarker() == null) {
-            item.setMarker(mFocusMarker);
-            markerCount++;
-            final AtomicInteger rotValue = new AtomicInteger(0);
-            if (timer != null) timer.cancel();
-            timer = new Timer();
-            TimerTask timerTask = new TimerTask() {
-                @Override
-                public void run() {
-                    float value = (float) (rotValue.incrementAndGet() * 10);
-                    item.setRotation(value);
-                    if (rotValue.get() > 36) rotValue.set(0);
-                    markerLayer.update();
-                    mMap.updateMap(true);
-                }
-            };
-            timer.schedule(timerTask, 300, 300);
-        } else {
-            item.setMarker(null);
-            markerCount--;
-            if (timer != null && markerCount == 0) timer.cancel();
-        }
+    public boolean onItemLongPress(int index, MarkerItem item) {
+        itemEvent(item);
 
         Toast.makeText(this, "Marker long press\n" + item.getTitle(), Toast.LENGTH_SHORT).show();
         return true;
-    }
-
-    protected class MapEventsReceiver extends Layer implements GestureListener {
-
-        MapEventsReceiver(Map map) {
-            super(map);
-        }
-
-        @Override
-        public boolean onGesture(Gesture g, MotionEvent e) {
-            if (g instanceof Gesture.Tap) {
-                GeoPoint p = mMap.viewport().fromScreenPoint(e.getX(), e.getY());
-                Toast.makeText(RotateMarkerOverlayActivity.this, "Map tap\n" + p, Toast.LENGTH_SHORT).show();
-                return true;
-            }
-            if (g instanceof Gesture.LongPress) {
-                GeoPoint p = mMap.viewport().fromScreenPoint(e.getX(), e.getY());
-                Toast.makeText(RotateMarkerOverlayActivity.this, "Map long press\n" + p, Toast.LENGTH_SHORT).show();
-                return true;
-            }
-            return false;
-        }
     }
 }