+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## New since 0.7.0
 
+- Symbol rotation [#294](https://github.com/mapsforge/vtm/issues/294)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.8.0)